+++ src/web/assets/cp/src/js/Craft.js
@@ -2049,6 +2049,68 @@ $.extend(Craft,
         setFocusWithin: function(container) {
             Garnish.setFocusWithin(container);
         },
+
+        /**
+         * Sets/removes attributes on an element.
+         *
+         * Attributes set to `null` or `false` will be removed.
+         *
+         * @param element
+         * @param {Object} attributes
+         */
+        setElementAttributes: function(element, attributes) {
+            const $element = $(element);
+
+            for (let name in attributes) {
+                if (!attributes.hasOwnProperty(name)) {
+                    continue;
+                }
+
+                let value = attributes[name];
+
+                if (value === null || value === false) {
+                    $element.removeAttr(name);
+                } else if (value === true) {
+                    $element.attr(name, '');
+                } else if ($.isArray(value) || $.isPlainObject(value)) {
+                    if (Craft.dataAttributes.includes(name)) {
+                        // Make sure it's an object
+                        value = Object.assign({}, value);
+                        for (let n in value) {
+                            if (!value.hasOwnProperty(n)) {
+                                continue;
+                            }
+                            let subValue = value[n];
+                            if (subValue === null || subValue === false) {
+                                continue;
+                            }
+                            if ($.isPlainObject(subValue) || $.isArray(subValue)) {
+                                subValue = JSON.stringify(subValue)
+                            } else if (subValue === true) {
+                                subValue = '';
+                            } else {
+                                subValue = this.escapeHtml(subValue);
+                            }
+                            $element.attr(`${name}-${n}`, subValue);
+                        }
+                    } else if (name === 'class') {
+                        // Make sure it's an array
+                        if ($.isPlainObject(value)) {
+                            value = Object.values(value);
+                        }
+                        for (let c of value) {
+                            $element.addClass(c);
+                        }
+                    } else if (name === 'style') {
+                        $element.css(value);
+                    } else {
+                        $element.attr(name, JSON.stringify(value));
+                    }
+                } else {
+                    $element.attr(name, this.escapeHtml(value));
+                }
+            }
+        },
     });
 
 // -------------------------------------------